requirements.txt
openai>=1.0.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
typing-extensions>=4.0.0
requests>=2.0.0
tqdm>=4.65.0  # for progress bars
logging>=0.5.0 